feat(web): Added booking description Modal
